-All models which are dependent on another are now defined within one file do realize SELECTS with nested includes -Realized an LocationFinder based on Google Geolocation API
