+++ Backend/model/FlightComment.js
@@ -1,18 +1,18 @@
-const { Sequelize, DataTypes } = require("sequelize");
-const db = require("../config/postgres.js");
+// const { Sequelize, DataTypes } = require("sequelize");
+// const db = require("../config/postgres.js");
 
-const FlightComment = db.sequelize.define("FlightComment", {
-  id: {
-    type: Sequelize.UUID,
-    defaultValue: Sequelize.UUIDV4,
-    allowNull: false,
-    primaryKey: true,
-  },
+// const FlightComment = db.sequelize.define("FlightComment", {
+//   id: {
+//     type: Sequelize.UUID,
+//     defaultValue: Sequelize.UUIDV4,
+//     allowNull: false,
+//     primaryKey: true,
+//   },
 
-  message: {
-    type: DataTypes.STRING,
-    allowNull: false,
-  },
-});
+//   message: {
+//     type: DataTypes.STRING,
+//     allowNull: false,
+//   },
+// });
 
-module.exports = FlightComment;
+// module.exports = FlightComment;
